+++ apps/vs-code-designer-react/src/app/app.tsx
@@ -2,15 +2,19 @@ import { createFileSystemConnection } from '../state/DesignerSlice';
 import type { AppDispatch, RootState } from '../state/Store';
 import { VSCodeContext } from '../webviewCommunication';
 import { DesignerCommandBar } from './DesignerCommandBar';
+import './app.less';
 import { getDesignerServices } from './servicesHelper';
 import { convertConnectionsDataToReferences } from './utilities/workflow';
+import { Spinner, SpinnerSize, Text } from '@fluentui/react';
 import type { ConnectionCreationInfo } from '@microsoft/designer-client-services-logic-apps';
 import type { ConnectionReferences } from '@microsoft/logic-apps-designer';
 import { DesignerProvider, BJSWorkflowProvider, Designer, getTheme, useThemeObserver } from '@microsoft/logic-apps-designer';
 import { isEmptyString, Theme } from '@microsoft/utils-logic-apps';
-import type { FileSystemConnectionInfo } from '@microsoft/vscode-extension';
+import type { FileSystemConnectionInfo, StandardApp } from '@microsoft/vscode-extension';
 import { ExtensionCommand } from '@microsoft/vscode-extension';
 import { useContext, useMemo, useState, useEffect } from 'react';
+import { useIntl } from 'react-intl';
+import { useQuery } from 'react-query';
 import { useDispatch, useSelector } from 'react-redux';
 
 export const App = () => {
@@ -30,9 +34,21 @@ export const App = () => {
     isMonitoringView,
     runId,
   } = vscodeState;
-  const [standardApp, setStandardApp] = useState(panelMetaData?.standardApp);
+  const [standardApp, setStandardApp] = useState<StandardApp | undefined>(panelMetaData?.standardApp);
   const [runInstance, setRunInstance] = useState<LogicAppsV2.RunInstanceDefinition | null>(null);
   const [theme, setTheme] = useState<Theme>(getTheme(document.body));
+  const intl = useIntl();
+
+  const intlText = {
+    ERROR_APP: intl.formatMessage({
+      defaultMessage: 'Something went wrong',
+      description: 'Something went wrong text',
+    }),
+    LOADING_APP: intl.formatMessage({
+      defaultMessage: 'Loading designer',
+      description: 'Loading designer text',
+    }),
+  };
 
   useThemeObserver(document.body, theme, setTheme, {
     attributes: true,
@@ -70,24 +86,67 @@ export const App = () => {
     return convertConnectionsDataToReferences(connectionData);
   }, [connectionData]);
 
-  useEffect(() => {
-    async function getRunInstance() {
-      if (isMonitoringView && !isEmptyString(runId)) {
-        const runServiceInstance = (await services.runService.getRun(runId)) as any;
-        const standardAppInstance = {
-          definition: runServiceInstance.properties.workflow.properties.definition,
-          kind: '',
-        };
-        setRunInstance(runServiceInstance);
-        setStandardApp(standardAppInstance);
-      }
+  const getRunInstance = () => {
+    if (isMonitoringView && !isEmptyString(runId) && panelMetaData !== null) {
+      return services.runService.getRun(runId);
     }
-    getRunInstance();
-  }, [isMonitoringView, runId, services]);
+    return;
+  };
+
+  const onRunInstanceSuccess = async (runDefinition: LogicAppsV2.RunInstanceDefinition) => {
+    const standardAppInstance = {
+      ...standardApp,
+      definition: runDefinition.properties.workflow.properties.definition,
+    } as StandardApp;
+    setRunInstance(runDefinition);
+    setStandardApp(standardAppInstance);
+  };
+
+  const onRunInstanceError = async () => {
+    setRunInstance(null);
+    setStandardApp(undefined);
+  };
+
+  const { refetch, isError, isFetching, isLoading, isRefetching } = useQuery<any>(['runInstance'], getRunInstance, {
+    refetchOnWindowFocus: false,
+    initialData: null,
+    onSuccess: onRunInstanceSuccess,
+    onError: onRunInstanceError,
+  });
+
+  useEffect(() => {
+    refetch();
+  }, [isMonitoringView, runId, services, refetch]);
 
   useEffect(() => {
     setStandardApp(panelMetaData?.standardApp);
   }, [panelMetaData]);
+
+  const errorApp = (
+    <Text className="designer--error" variant="xLarge" block>
+      {intlText.ERROR_APP}{' '}
+    </Text>
+  );
+
+  const loadingApp = <Spinner className="designer--loading" size={SpinnerSize.large} label={intlText.LOADING_APP} />;
+
+  const designerCommandBar =
+    readOnly && !isMonitoringView ? null : (
+      <DesignerCommandBar
+        isMonitoringView={isMonitoringView}
+        isDisabled={isError || isFetching || isLoading}
+        isRefreshing={isRefetching}
+        onRefresh={refetch}
+      />
+    );
+
+  const designerApp = standardApp ? (
+    <BJSWorkflowProvider workflow={{ definition: standardApp.definition, connectionReferences }} runInstance={runInstance}>
+      <Designer />
+    </BJSWorkflowProvider>
+  ) : (
+    errorApp
+  );
 
   return (
     <DesignerProvider
@@ -99,12 +158,8 @@ export const App = () => {
         services: services,
       }}
     >
-      {standardApp ? (
-        <BJSWorkflowProvider workflow={{ definition: standardApp.definition, connectionReferences }} runInstance={runInstance}>
-          {readOnly && !isMonitoringView ? null : <DesignerCommandBar isMonitoringView={isMonitoringView} />}
-          <Designer />
-        </BJSWorkflowProvider>
-      ) : null}
+      {designerCommandBar}
+      {isError ? errorApp : isFetching || isLoading ? loadingApp : designerApp}
     </DesignerProvider>
   );
 };