+++ libs/designer-ui/src/lib/authentication/MSIAuth/MSIAuth.tsx
@@ -1,5 +1,5 @@
 import type { AuthProps, MSIProps } from '..';
-import type { ChangeState, TokenPickerHandler } from '../../editor/base';
+import type { ChangeState, GetTokenPickerHandler } from '../../editor/base';
 import { AuthenticationDropdown } from '../AuthenticationDropdown';
 import { AuthenticationProperty } from '../AuthenticationProperty';
 import { AUTHENTICATION_PROPERTIES, containsUserAssignedIdentities } from '../util';
@@ -15,7 +15,7 @@ import { useIntl } from 'react-intl';
 interface MSIAuthenticationProps {
   msiProps: MSIProps;
   identity?: ManagedIdentity;
-  tokenPickerHandler: TokenPickerHandler;
+  getTokenPicker: GetTokenPickerHandler;
   onManagedIdentityChange(event: React.FormEvent<HTMLDivElement>, item: IDropdownOption): void;
   setCurrentProps: Dispatch<SetStateAction<AuthProps>>;
 }
@@ -23,7 +23,7 @@ interface MSIAuthenticationProps {
 export const MSIAuthentication = ({
   identity,
   msiProps,
-  tokenPickerHandler,
+  getTokenPicker,
   onManagedIdentityChange,
   setCurrentProps,
 }: MSIAuthenticationProps): JSX.Element => {
@@ -65,14 +65,14 @@ export const MSIAuthentication = ({
           <AuthenticationProperty
             initialValue={msiAudience}
             AuthProperty={AUTHENTICATION_PROPERTIES.MSI_AUDIENCE}
-            tokenPickerHandler={tokenPickerHandler}
+            getTokenPicker={getTokenPicker}
             onBlur={updateMsiAudience}
           />
         </>
       ) : (
         <MSIAuthenticationDefault
           msiProps={msiProps}
-          tokenPickerHandler={tokenPickerHandler}
+          getTokenPicker={getTokenPicker}
           onManagedIdentityChange={onManagedIdentityChange}
           onBlur={updateMsiAudience}
         />