+++ libs/designer-ui/src/lib/tokenpicker/tokenpickersection/tokenpickernomatches.tsx
@@ -1,10 +1,7 @@
 import { AnnouncedMatches } from '../../announcedmatches';
 import { useIntl } from 'react-intl';
 
-interface TokenPickerNoMatchesProps {
-  noItems: boolean;
-}
-export const TokenPickerNoMatches = ({ noItems }: TokenPickerNoMatchesProps): JSX.Element => {
+export const TokenPickerNoMatches = (): JSX.Element => {
   const intl = useIntl();
   const noMatchesText = intl.formatMessage({
     defaultMessage: 'There are no results for your search',
@@ -24,18 +21,14 @@ export const TokenPickerNoMatches = ({ noItems }: TokenPickerNoMatchesProps): JS
   });
   return (
     <>
-      {noItems ? (
-        <>
-          <div className="msla-token-picker-no-matches">
-            <div className="msla-token-picker-no-results">{noMatchesText}</div>
-            <hr />
-            <div className="msla-token-picker-no-matches-header">{searchTipsHeader}</div>
-            <div className="msla-token-picker-no-matches-tip">{searchTip1}</div>
-            <div className="msla-token-picker-no-matches-tip">{searchTip2}</div>
-          </div>
-          <AnnouncedMatches count={0} visible={true} />
-        </>
-      ) : null}
+      <div className="msla-token-picker-no-matches">
+        <div className="msla-token-picker-no-results">{noMatchesText}</div>
+        <hr />
+        <div className="msla-token-picker-no-matches-header">{searchTipsHeader}</div>
+        <div className="msla-token-picker-no-matches-tip">{searchTip1}</div>
+        <div className="msla-token-picker-no-matches-tip">{searchTip2}</div>
+      </div>
+      <AnnouncedMatches count={0} visible={true} />
     </>
   );
 };