+++ libs/designer/src/lib/core/actions/bjsworkflow/initialize.ts
@@ -13,6 +13,8 @@ import { initializeParameters } from '../../state/workflowparameters/workflowpar
 import type { RootState } from '../../store';
 import { getBrandColorFromConnector, getIconUriFromConnector } from '../../utils/card';
 import { getTriggerNodeId, isRootNodeInGraph } from '../../utils/graph';
+import type { OpenApiConnectionSerializedInputs } from '../../utils/openapi/inputsbuilder';
+import { OpenApiOperationInputsBuilder } from '../../utils/openapi/inputsbuilder';
 import { getSplitOnOptions, getUpdatedManifestForSchemaDependency, getUpdatedManifestForSpiltOn, toOutputInfo } from '../../utils/outputs';
 import {
   addRecurrenceParametersInGroup,
@@ -29,6 +31,7 @@ import {
 import { createLiteralValueSegment } from '../../utils/parameters/segment';
 import { getOutputParametersFromSwagger } from '../../utils/swagger/operation';
 import { convertOutputsToTokens, getBuiltInTokens, getTokenNodeIds } from '../../utils/tokens';
+import { isOpenApiConnectionType } from './connections';
 import type { NodeInputsWithDependencies, NodeOutputsWithDependencies } from './operationdeserializer';
 import type { Settings } from './settings';
 import type {
@@ -97,37 +100,47 @@ export const getInputParametersFromManifest = (
   let primaryInputParametersInArray = unmap(primaryInputParameters);
 
   if (stepDefinition) {
-    const { inputsLocation } = manifest.properties;
-    const operationData = clone(stepDefinition);
+    const nodeType: string | undefined = stepDefinition.type;
+    if (nodeType && isOpenApiConnectionType(nodeType)) {
+      const stepDefinitionInputs: OpenApiConnectionSerializedInputs | undefined = stepDefinition.inputs;
 
-    // In the case of retry policy, it is treated as an input
-    // avoid pushing a parameter for it as it is already being
-    // handled in the settings store.
-    // NOTE: this could be expanded to more settings that are treated as inputs.
-    if (
-      manifest.properties.settings &&
-      manifest.properties.settings.retryPolicy &&
-      operationData.inputs &&
-      operationData.inputs[PropertyName.RETRYPOLICY]
-    ) {
-      delete operationData.inputs.retryPolicy;
+      primaryInputParametersInArray = new OpenApiOperationInputsBuilder().loadStaticInputValuesFromDefinition(
+        stepDefinitionInputs,
+        primaryInputParametersInArray
+      );
+    } else {
+      const { inputsLocation } = manifest.properties;
+      const operationData = clone(stepDefinition);
+
+      // In the case of retry policy, it is treated as an input
+      // avoid pushing a parameter for it as it is already being
+      // handled in the settings store.
+      // NOTE: this could be expanded to more settings that are treated as inputs.
+      if (
+        manifest.properties.settings &&
+        manifest.properties.settings.retryPolicy &&
+        operationData.inputs &&
+        operationData.inputs[PropertyName.RETRYPOLICY]
+      ) {
+        delete operationData.inputs.retryPolicy;
+      }
+
+      if (
+        manifest.properties.connectionReference &&
+        manifest.properties.connectionReference.referenceKeyFormat === ConnectionReferenceKeyFormat.Function
+      ) {
+        delete operationData.inputs.function;
+      }
+
+      primaryInputParametersInArray = updateParameterWithValues(
+        'inputs.$',
+        getInputsValueFromDefinitionForManifest(inputsLocation ?? ['inputs'], manifest, operationData, primaryInputParametersInArray),
+        '',
+        primaryInputParametersInArray,
+        (!inputsLocation || !!inputsLocation.length) && !manifest.properties.inputsLocationSwapMap /* createInvisibleParameter */,
+        false /* useDefault */
+      );
     }
-
-    if (
-      manifest.properties.connectionReference &&
-      manifest.properties.connectionReference.referenceKeyFormat === ConnectionReferenceKeyFormat.Function
-    ) {
-      delete operationData.inputs.function;
-    }
-
-    primaryInputParametersInArray = updateParameterWithValues(
-      'inputs.$',
-      getInputsValueFromDefinitionForManifest(inputsLocation ?? ['inputs'], manifest, operationData, primaryInputParametersInArray),
-      '',
-      primaryInputParametersInArray,
-      (!inputsLocation || !!inputsLocation.length) && !manifest.properties.inputsLocationSwapMap /* createInvisibleParameter */,
-      false /* useDefault */
-    );
   } else {
     loadParameterValuesArrayFromDefault(primaryInputParametersInArray);
   }