+++ libs/designer/src/lib/ui/panel/panelTabs/scratchTab.tsx
@@ -1,8 +1,10 @@
 import constants from '../../../common/constants';
 import type { TokenGroup } from '../../../core/utils/tokens';
 import { getExpressionTokenSections } from '../../../core/utils/tokens';
-import type { OutputToken, PanelTab, TokenPickerHandler, ValueSegment } from '@microsoft/designer-ui';
+import type { OutputToken, PanelTab, TokenPickerMode, ValueSegment } from '@microsoft/designer-ui';
 import {
+  AuthenticationType,
+  Combobox,
   RowDropdownOptions,
   GroupType,
   ArrayEditor,
@@ -12,13 +14,15 @@ import {
   // GroupDropdownOptions,
   // QueryBuilderEditor, // DictionaryType, // EditorLanguage,
   UntilEditor,
-  ValueSegmentType, // CodeEditor, // HTMLEditor,
+  ValueSegmentType,
+  CodeEditor,
+  EditorLanguage, // HTMLEditor,
   TokenPicker, // TokenType,
   // DictionaryEditor,
   testTokenSegment, // SchemaEditor, // Combobox,
   // ArrayEditor,
-  Scratch, // StringEditor, // AuthenticationEditor,
-  // DropdownEditor,
+  Scratch, // StringEditor,
+  AuthenticationEditor, // DropdownEditor,
   outputToken,
   outputToken2, // RowDropdownOptions,
 } from '@microsoft/designer-ui';
@@ -64,9 +68,10 @@ export const ScratchTab = () => {
   const GetTokenPicker = (
     editorId: string,
     labelId: string,
+    tokenPickerMode?: TokenPickerMode,
+    closeTokenPicker?: () => void,
     onClick?: (b: boolean) => void,
-    tokenClicked?: (token: ValueSegment) => void,
-    tokenPickerHide?: () => void
+    tokenClicked?: (token: ValueSegment) => void
   ): JSX.Element => {
     return (
       <TokenPicker
@@ -74,20 +79,20 @@ export const ScratchTab = () => {
         labelId={labelId}
         tokenGroup={testTokenGroup}
         expressionGroup={expressionGroup}
+        initialMode={tokenPickerMode}
+        closeTokenPicker={closeTokenPicker}
         tokenPickerFocused={onClick}
         getValueSegmentFromToken={getValueSegmentFromToken}
         tokenClickedCallback={tokenClicked}
-        tokenPickerHide={tokenPickerHide}
       />
     );
   };
 
-  const tokenPickerHandler: TokenPickerHandler = { getTokenPicker: GetTokenPicker, tokenPickerProps: {} };
   const children = (): React.ReactNode => {
     return (
       <>
         {/* <QueryBuilderEditor
-          tokenPickerHandler={tokenPickerHandler}
+getTokenPicker={GetTokenPicker}
           readonly={false}
           groupProps={{
             type: GroupType.GROUP,
@@ -112,7 +117,7 @@ export const ScratchTab = () => {
             ],
           }}
         /> */}
-        {/* <CodeEditor
+        <CodeEditor
           initialValue={[
             {
               type: ValueSegmentType.LITERAL,
@@ -134,9 +139,9 @@ export const ScratchTab = () => {
               id: guid(),
             },
           ]}
-          tokenPickerHandler={tokenPickerHandler}
+          getTokenPicker={GetTokenPicker}
           language={EditorLanguage.javascript}
-        /> */}
+        />
         {/* <DropdownEditor
           multiSelect={true}
           initialValue={[
@@ -151,7 +156,22 @@ export const ScratchTab = () => {
             { displayName: 'DELETE', value: 'DELETE', key: 'DELETE', disabled: false },
           ]}
         /> */}
-        {/* <AuthenticationEditor initialValue={[]} tokenPickerHandler={tokenPickerHandler} AuthenticationEditorOptions={{}} authProps={{}} /> */}
+        <AuthenticationEditor
+          initialValue={[]}
+          getTokenPicker={GetTokenPicker}
+          options={{
+            supportedAuthTypes: [
+              AuthenticationType.BASIC,
+              AuthenticationType.NONE,
+              AuthenticationType.CERTIFICATE,
+              AuthenticationType.MSI,
+              AuthenticationType.OAUTH,
+              AuthenticationType.RAW,
+            ],
+          }}
+          type={AuthenticationType.BASIC}
+          authenticationValue={{}}
+        />
         <ArrayEditor
           type={ArrayType.SIMPLE}
           labelProps={{ text: 'Input Array', isRequiredField: true }}
@@ -166,7 +186,7 @@ export const ScratchTab = () => {
             { id: guid(), type: ValueSegmentType.LITERAL, value: 'More Text' },
             { id: guid(), type: ValueSegmentType.LITERAL, value: '"\n]' },
           ]}
-          tokenPickerHandler={tokenPickerHandler}
+          getTokenPicker={GetTokenPicker}
         />
 
         <ArrayEditor
@@ -185,10 +205,10 @@ export const ScratchTab = () => {
             testTokenSegment,
             { id: guid(), type: ValueSegmentType.LITERAL, value: '",\n            "p2": ""\n        }\n    }\n]' },
           ]}
-          tokenPickerHandler={tokenPickerHandler}
+          getTokenPicker={GetTokenPicker}
         />
 
-        {/* <Combobox
+        <Combobox
           options={[
             { displayName: 'GET', value: 'GET', key: 'GET', disabled: false },
             { displayName: 'PUT', value: 'PUT', key: 'PUT', disabled: false },
@@ -198,9 +218,9 @@ export const ScratchTab = () => {
           ]}
           placeholder="Method is Required"
           initialValue={[{ id: '0', type: ValueSegmentType.LITERAL, value: 'PUT' }]}
-          tokenPickerHandler={tokenPickerHandler}
+          getTokenPicker={GetTokenPicker}
           // readonly={true}
-        /> */}
+        />
 
         {/* <SchemaEditor
           initialValue={[{ id: '0', type: ValueSegmentType.LITERAL, value: '{\n"type": "object",\n"properties" : {}\n}' }]}
@@ -271,7 +291,7 @@ export const ScratchTab = () => {
             { id: guid(), type: ValueSegmentType.LITERAL, value: 'Value2 Text' },
             { id: guid(), type: ValueSegmentType.LITERAL, value: '"\n}' },
           ]}
-          tokenPickerHandler={tokenPickerHandler}
+          getTokenPicker={GetTokenPicker}
         /> */}
         {/* <DictionaryEditor
           initialItems={[
@@ -338,12 +358,12 @@ export const ScratchTab = () => {
             { id: guid(), type: ValueSegmentType.LITERAL, value: 'Value2 Text' },
             { id: guid(), type: ValueSegmentType.LITERAL, value: '"\n}' },
           ]}
-          tokenPickerHandler={tokenPickerHandler}
+          getTokenPicker={GetTokenPicker}
           dictionaryType={DictionaryType.TABLE}
           keyTitle={'Header'}
           valueTitle={'Value'}
         /> */}
-        {/* <DictionaryEditor initialValue={[testTokenSegment]} tokenPickerHandler={tokenPickerHandler} /> */}
+        {/* <DictionaryEditor initialValue={[testTokenSegment]}           getTokenPicker={GetTokenPicker}*/}
         {/* <StringEditor
           initialValue={[
             testTokenSegment,
@@ -351,12 +371,12 @@ export const ScratchTab = () => {
             testTokenSegment,
             testTokenSegment,
           ]}
-          tokenPickerHandler={tokenPickerHandler}
+          getTokenPicker={GetTokenPicker}
         /> */}
-        {/* <HTMLEditor initialValue={[]} placeholder="Specify the body of the mail" tokenPickerHandler={tokenPickerHandler} /> */}
+        {/* <HTMLEditor initialValue={[]} placeholder="Specify the body of the mail"           getTokenPicker={GetTokenPicker}*/}
         {
           <UntilEditor
-            tokenPickerHandler={tokenPickerHandler}
+            getTokenPicker={GetTokenPicker}
             items={{ type: GroupType.ROW, operand1: [], operand2: [], operator: RowDropdownOptions.EQUALS }}
           />
         }