+++ libs/services/designer-client-services/src/lib/base/operationmanifest.ts
@@ -107,6 +107,8 @@ const appendtoarrayvariable = 'appendtoarrayvariable';
 const appendtostringvariable = 'appendtostringvariable';
 const batch = 'batch';
 const sendtobatch = 'sendtobatch';
+const xslttransform = 'xslttransform';
+const datamapper = 'datamapper';
 
 export const apiManagementConnectorId = 'connectionProviders/apiManagementOperation';
 export const azureFunctionConnectorId = 'connectionProviders/azureFunctionOperation';
@@ -121,6 +123,8 @@ const scheduleConnectorId = 'connectionProviders/schedule';
 const httpConnectorId = 'connectionProviders/http';
 const variableConnectorId = 'connectionProviders/variable';
 const rosettanetConnectorId = 'connectionProviders/rosettaNetOperations';
+const liquidConnectorId = 'connectionProviders/liquidOperations';
+const dataMapperConnectorId = 'connectionProviders/dataMapperOperations';
 
 const azurefunction = 'azureFunction';
 const appservice = 'appservice';
@@ -176,6 +180,7 @@ export const supportedBaseManifestTypes = [
   terminate,
   until,
   wait,
+  xslttransform,
 ];
 
 export type getAccessTokenType = () => Promise<string>;
@@ -263,6 +268,7 @@ export function isBuiltInOperation(definition: any): boolean {
     case terminate:
     case until:
     case wait:
+    case xslttransform:
       return true;
 
     case appservice:
@@ -276,8 +282,8 @@ export function isBuiltInOperation(definition: any): boolean {
 }
 
 export function getBuiltInOperationInfo(definition: any, isTrigger: boolean): OperationInfo {
-  const normalizedOperationType = definition.type.toLowerCase();
-  const kind = definition.kind ? definition.kind.toLowerCase() : undefined;
+  const normalizedOperationType = definition.type?.toLowerCase();
+  const kind = definition.kind?.toLowerCase();
 
   if (kind === undefined) {
     const operationInfo = builtInOperationsMetadata[normalizedOperationType];
@@ -286,7 +292,6 @@ export function getBuiltInOperationInfo(definition: any, isTrigger: boolean): Op
     }
   }
 
-  const liquidConnectorId = 'connectionProviders/liquidOperations';
   switch (normalizedOperationType) {
     case expression:
       switch (definition.kind?.toLowerCase()) {
@@ -430,6 +435,17 @@ export function getBuiltInOperationInfo(definition: any, isTrigger: boolean): Op
         connectorId: workflowConnectorId,
         operationId: invokeworkflow,
       };
+
+    case xslttransform:
+      switch (kind) {
+        case datamapper:
+          return {
+            connectorId: dataMapperConnectorId,
+            operationId: xslttransform,
+          };
+        default:
+          throw new UnsupportedException(`Unsupported operation kind ${kind}`);
+      }
 
     default:
       throw new UnsupportedException(`Unsupported built in operation type ${normalizedOperationType}`);
@@ -596,6 +612,10 @@ const builtInOperationsMetadata: Record<string, OperationInfo> = {
   [rosettanetwaitforresponse]: {
     connectorId: rosettanetConnectorId,
     operationId: 'rosettaNetWaitForResponse',
+  },
+  [xslttransform]: {
+    connectorId: 'connectionProviders/dataMapperOperations',
+    operationId: 'xsltTransform',
   },
 };